Union de mapa de rutas y puntos con infoview personalizada!
