+++ app/src/main/AndroidManifest.xml
@@ -41,16 +41,6 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/google_maps_key" />
 
-        <activity
-            android:name=".MapsActivity"
-            android:label="@string/title_activity_maps">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.example.lupe.siin.MainActivity" />
-        </activity>
-        <activity
-            android:name=".ShapesActivity"
-            android:label="@string/title_activity_shapes" />
         <activity android:name=".MapFragmentActivity" />
         <activity android:name=".FichaActivity" />
         <activity android:name=".SeguimientoActivity" />
