+++ app/src/main/res/layout/activity_map_fragment.xml
