--- app/src/main/res/layout/activity_sample_with_map_view.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_sample_with_map_view"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.example.lupe.siin.MapViewActivity">
-
-    <com.appolica.interactiveinfowindow.customview.TouchInterceptFrameLayout
-        android:id="@+id/mapViewContainer"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content">
-
-        <com.google.android.gms.maps.MapView
-            android:id="@+id/mapView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"/>
-
-    </com.appolica.interactiveinfowindow.customview.TouchInterceptFrameLayout>
-
-</RelativeLayout>
